Refactor "mtl" -> "end_model"
