+++ snorkel/end_model/schedulers/sequential_scheduler.py
@@ -1,4 +1,4 @@
-from snorkel.mtl.schedulers.scheduler import Scheduler
+from snorkel.end_model.schedulers.scheduler import Scheduler
 
 
 class SequentialScheduler(Scheduler):
