+++ snorkel/end_model/scorer.py
