+++ snorkel/end_model/trainer.py
@@ -6,12 +6,12 @@ import torch
 import torch.optim as optim
 from tqdm import tqdm
 
-from snorkel.mtl.loggers import Checkpointer, LogManager, LogWriter, TensorBoardWriter
-from snorkel.mtl.model import MultitaskModel
-from snorkel.mtl.schedulers.sequential_scheduler import SequentialScheduler
-from snorkel.mtl.schedulers.shuffled_scheduler import ShuffledScheduler
-from snorkel.mtl.snorkel_config import default_config
-from snorkel.mtl.utils import recursive_merge_dicts
+from snorkel.end_model.loggers import Checkpointer, LogManager, LogWriter, TensorBoardWriter
+from snorkel.end_model.model import MultitaskModel
+from snorkel.end_model.schedulers.sequential_scheduler import SequentialScheduler
+from snorkel.end_model.schedulers.shuffled_scheduler import ShuffledScheduler
+from snorkel.end_model.snorkel_config import default_config
+from snorkel.end_model.utils import recursive_merge_dicts
 
 
 class Trainer(object):
