+++ test/end_model/loggers/test_log_manager.py
@@ -2,8 +2,8 @@ import os
 import shutil
 import unittest
 
-from snorkel.mtl.loggers.checkpointer import Checkpointer
-from snorkel.mtl.loggers.log_manager import LogManager
+from snorkel.end_model.loggers.checkpointer import Checkpointer
+from snorkel.end_model.loggers.log_manager import LogManager
 
 TEST_LOG_DIR = "test/mtl/loggers/logs"
