+++ test/end_model/test_mtl_utils.py
@@ -2,7 +2,7 @@ import unittest
 
 import torch
 
-from snorkel.mtl.utils import list_to_tensor, pad_batch
+from snorkel.end_model.utils import list_to_tensor, pad_batch
 
 
 class UtilsTest(unittest.TestCase):
