+++ test/end_model/test_trainer.py
@@ -6,12 +6,12 @@ import torch
 import torch.nn as nn
 
 from snorkel.model.utils import set_seed
-from snorkel.mtl.data import MultitaskDataLoader, MultitaskDataset
-from snorkel.mtl.model import MultitaskModel
-from snorkel.mtl.modules.utils import ce_loss, softmax
-from snorkel.mtl.scorer import Scorer
-from snorkel.mtl.task import Operation, Task
-from snorkel.mtl.trainer import Trainer
+from snorkel.end_model.data import MultitaskDataLoader, MultitaskDataset
+from snorkel.end_model.model import MultitaskModel
+from snorkel.end_model.modules.utils import ce_loss, softmax
+from snorkel.end_model.scorer import Scorer
+from snorkel.end_model.task import Operation, Task
+from snorkel.end_model.trainer import Trainer
 
 SEED = 123
